backend/pyproject.toml
[project]
name = "neurocareer-backend"
version = "0.1.0"
description = "FastAPI + PydanticAI service for the NeuroCareer AI mentor"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "pydantic-ai>=0.0.13",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "apscheduler>=3.10.4",
    "python-multipart>=0.0.9",
    "pypdf>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
    "hindsight-client>=0.1.2",
    "structlog>=24.1.0",
    "bcrypt>=4.0.0",
    "PyJWT>=2.8.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
